chore: changed weight to double in donation model

+++ lib/model/donation.dart
@@ -2,7 +2,7 @@ class Donation {
   String id;
   String donor;
   String category;
-  int weight;
+  double weight;
   String address;
   String contactNo;
   String pickUpDateTime;
@@ -10,16 +10,15 @@ class Donation {
   String? photo;
   int status;
 
-  Donation({
-    required this.id,
-    required this.donor, 
-    required this.category, 
-    required this.weight, 
-    required this.address, 
-    required this.contactNo,
-    required this.pickUpDateTime, 
-    required this.dropOffDateTime,
-    this.photo,
-    required this.status
-  });
+  Donation(
+      {required this.id,
+      required this.donor,
+      required this.category,
+      required this.weight,
+      required this.address,
+      required this.contactNo,
+      required this.pickUpDateTime,
+      required this.dropOffDateTime,
+      this.photo,
+      required this.status});
 }
